Add option 2 for getting weather forecast by ZIP code
This commit introduces the functionality to retrieve weather forecasts for multiple ZIP codes. The user is prompted to enter ZIP code(s) and country code(s) separated by commas. The code validates the input, retrieves coordinates for each ZIP code, and then attempts to fetch the corresponding weather information.

Comments have been added to explain the logic of the code. It includes sections for handling various scenarios, such as when no ZIP codes are provided and when there are multiple ZIP codes to process. The comments provide a clear understanding of the purpose and flow of the code.